# engine/openapi/cz.solvina.options.spreads.yaml
openapi: 3.0.3
info:
  title: Options Engine - Spreads API
  description: Bull Put Spread positions and scanner management
  version: 1.0.0

paths:
  /spreads:
    get:
      tags:
        - Spreads
      summary: List spread positions
      operationId: listSpreads
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, CLOSING, CLOSED_PROFIT, CLOSED_STOP, CLOSED_TIME, CLOSED_MANUAL]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 25
      responses:
        '200':
          description: Paginated spread positions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedSpreadsDto'

  /spreads/{id}:
    get:
      tags:
        - Spreads
      summary: Get a spread position by ID
      operationId: getSpreadById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Spread position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpreadDto'
        '404':
          description: Not found

  /spreads/{id}/close:
    post:
      tags:
        - Spreads
      summary: Soft-close a spread at mid price (limit orders)
      operationId: softCloseSpread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Spread closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpreadDto'
        '404':
          description: Spread not found
        '409':
          description: Spread is not open

  /spreads/{id}/refresh-pnl:
    post:
      tags:
        - Spreads
      summary: Recompute and persist current P&L for a spread (ad-hoc, hits market data)
      operationId: refreshSpreadPnl
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated spread with fresh P&L
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpreadDto'
        '404':
          description: Spread not found

  /spreads/{id}/close-force:
    post:
      tags:
        - Spreads
      summary: Force-close a spread at market price
      operationId: forceCloseSpread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Spread closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpreadDto'
        '404':
          description: Spread not found
        '409':
          description: Spread is not open

  /scanner/run:
    post:
      tags:
        - Scanner
      summary: Trigger a manual scanner run
      operationId: triggerScan
      responses:
        '202':
          description: Scan accepted and running asynchronously

  /scanner/status:
    get:
      tags:
        - Scanner
      summary: Get scanner and monitor status including kill-switch state
      operationId: getScannerStatus
      responses:
        '200':
          description: Scanner status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScannerStatusDto'

  /scanner/pause:
    post:
      tags:
        - Scanner
      summary: Pause new-entry scanner (existing positions are unaffected)
      operationId: pauseScanner
      responses:
        '204':
          description: Scanner paused

  /scanner/resume:
    post:
      tags:
        - Scanner
      summary: Resume new-entry scanner
      operationId: resumeScanner
      responses:
        '204':
          description: Scanner resumed

  /monitor/pause:
    post:
      tags:
        - Scanner
      summary: Pause spread exit monitor (no automatic closes will be triggered)
      operationId: pauseMonitor
      responses:
        '204':
          description: Monitor paused

  /monitor/resume:
    post:
      tags:
        - Scanner
      summary: Resume spread exit monitor
      operationId: resumeMonitor
      responses:
        '204':
          description: Monitor resumed

components:
  schemas:
    SpreadDto:
      type: object
      required:
        - symbol
        - status
        - soldStrike
        - boughtStrike
        - expiryDate
        - creditPerShare
        - maxRiskPerShare
        - quantity
        - openedAt
      properties:
        id:
          type: string
          format: uuid
        symbol:
          type: string
        status:
          type: string
        soldStrike:
          type: number
        boughtStrike:
          type: number
        expiryDate:
          type: string
          format: date
        creditPerShare:
          type: number
        maxRiskPerShare:
          type: number
        quantity:
          type: integer
        ivRankAtEntry:
          type: number
          nullable: true
        underlyingPriceAtEntry:
          type: number
          nullable: true
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
        closeReason:
          type: string
          nullable: true
        closePricePerShare:
          type: number
          nullable: true
        currentSpreadValue:
          type: number
          nullable: true
        currentPnl:
          type: number
          nullable: true

    PagedSpreadsDto:
      type: object
      required: [content, totalElements, totalPages, page, size]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SpreadDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer

    ScannerStatusDto:
      type: object
      required:
        - openSpreadCount
        - ivRanks
        - scannerPaused
        - monitorPaused
      properties:
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        openSpreadCount:
          type: integer
          format: int64
        ivRanks:
          type: object
          additionalProperties:
            type: number
        scannerPaused:
          type: boolean
        monitorPaused:
          type: boolean